Move ExportWeatherButton to weather_viewer.py
